% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd_units.R
\name{as_sunits_rcrd}
\alias{as_sunits_rcrd}
\title{Coerce dataframes and lists to sunits}
\usage{
as_sunits_rcrd(x, value_from = "value", units_from = "units")
}
\arguments{
\item{x}{dataframe or list to coerce}
}
\description{
Coerce dataframes and lists to sunits
}
