% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert unitful quantities}
\usage{
convert(x, new_x, from, to)
}
\arguments{
\item{x}{unitful sunits}

\item{new_x}{replacement value}

\item{from}{length 1 character of unit to convert from}

\item{to}{length 1 character of unit to convert to}
}
\description{
Convert unitful quantities
}
