% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl_units.R
\name{new_sunits_dbl}
\alias{new_sunits_dbl}
\title{Simple Double w/ Units Class}
\usage{
new_sunits_dbl(value = double(), units = character())
}
\arguments{
\item{units}{character representing the units (may be mixed)}

\item{x}{a double}
}
\value{
record-style s3 `sunits` object
}
\description{
Simple Double w/ Units Class
}
