% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd_units.R, R/simple_units.R
\name{new_sunits_rcrd}
\alias{new_sunits_rcrd}
\title{Simple Record-Style Units Class}
\usage{
new_sunits_rcrd(value = double(), units = character())

new_sunits_rcrd(value = double(), units = character())
}
\arguments{
\item{units}{character representing the units (may be mixed)}

\item{x}{a double}
}
\value{
record-style s3 `sunits_rcrd` object

record-style s3 `sunits_rcrd` object
}
\description{
Simple Record-Style Units Class

Simple Double w/ Units Class
}
